pyproject.toml
[tool.mypy]
ignore_missing_imports = true

[project]
name = "advent-of-code"
version = "0.1.0"
description = "Advent of Code challenge solutions"
authors = []
requires-python = ">=3.12.0"
dependencies = [
    "ipython>=9.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.scripts]
aoc = "advent_of_code.runner:main"
aoc-scaffold = "advent_of_code.scaffold:main"

[dependency-groups]
dev = [
    "coverage[toml]>=6.5",
    "coveralls>=3.3.1",
    "mypy>=1.11.2",
    "py-spy>=0.3.14",
    "pytest>=8.3.3",
    "ruff>=0.6.5",
]


[[tool.uv.index]]
name = "hudl"
url = "https://hudlaml-690616407375.d.codeartifact.us-east-1.amazonaws.com/pypi/hudlaml/simple/"
explicit = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
